requirements.txt
asgiref==3.8.1
config==0.5.1
Django==5.0.7
djangorestframework==3.15.2
psycopg==3.2.1
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
